Added collectables
Added coins and dollars. They spin in place. When a coins is collected a "1" floats to show the points earned and a "4" for the dollar. The dollar is rare and it has 1 to 8 chances to coming out since they are worth more.
Note: we can add a sound to when the collectables are collected. I have the function already ready it's just commented out since i got no sound atm. changed the scoring so it could be the total amount of points instead of making it based on speed.